+++ chicken_dinner/models/telemetry/events.py
@@ -17,7 +17,8 @@ class TelemetryEvent(object):
     embedded telemetry objects recursively. Converts all event and object keys
     to snake-cased key names.
 
-    :param data: the JSON object data associated with the telemetry event
+    :param dict data: the JSON object data associated with the telemetry event
+    :param bool map_assets: whether to map asset ids to asset names
     """
 
     def __init__(self, data, map_assets=False):
@@ -66,13 +67,23 @@ class TelemetryEvent(object):
 
     def dumps(self):
         """Serialize the event to a JSON string."""
-        return json.dumps(
-            self,
-            default=lambda x: remove_from_dict(x.__dict__, ["reference"]),
-            sort_keys=True,
-            indent=4
-        )
+        return json.dumps(self, default=lambda x: remove_from_dict(x.__dict__, ["reference"]), sort_keys=True, indent=4)
 
     def to_dict(self):
         """Get the event as a dict."""
         return json.loads(self.dumps())
+
+    def keys(self):
+        """Get all attributes names."""
+        return [k for k in self.__dict__.keys() if k[0] != "_"]
+
+    def values(self):
+        """Get all attribute values."""
+        return [self.__dict__[k] for k in self.__dict__.keys() if k[0] != "_"]
+
+    def items(self):
+        """Iterate through the attributes dictionary."""
+        for k, v in self.__dict__.items():
+            if k[0] == "_":
+                continue
+            yield k, v