+++ chicken_dinner/util.py
@@ -3,11 +3,7 @@ import copy
 import re
 
 
-stats_map = {
-    "DBNOs": "dbnos",
-    "dBNOs": "dbnos",
-    "top10s": "top_10s",
-}
+stats_map = {"DBNOs": "dbnos", "dBNOs": "dbnos", "top10s": "top_10s", "dBNOId": "dbno_id"}
 
 
 def camel_to_snake(name):
@@ -15,8 +11,8 @@ def camel_to_snake(name):
         return stats_map[name]
     except KeyError as exc:
         pass
-    s1 = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', name)
-    return re.sub('([a-z0-9])([A-Z])', r'\1_\2', s1).lower()
+    s1 = re.sub("(.)([A-Z][a-z]+)", r"\1_\2", name)
+    return re.sub("([a-z0-9])([A-Z])", r"\1_\2", s1).lower()
 
 
 def remove_from_dict(d, keys):